+++ DemoTwitter/DemoTwitter.DataAccessLayer/Tweets/TweetsRepository.cs
@@ -7,22 +7,37 @@ namespace DemoTwitter.DataAccessLayer.Tweets
     {
         private readonly Twitter_dbEntities dbContext = new Twitter_dbEntities();
 
-        public void Add(Tweet tweet)
+        public bool Add(Tweet tweet)
         {
+            if (tweet == null)
+            {
+                return false;
+            }
             dbContext.Tweets.Add(tweet);
             dbContext.SaveChanges();
+            return true;
         }
 
-        public void Update(Tweet oldTweet, Tweet newTweet)
+        public bool Update(Tweet oldTweet, Tweet newTweet)
         {
+            if (newTweet == null)
+            {
+                return false;
+            }
             oldTweet.text = newTweet.text;
             oldTweet.post_date = newTweet.post_date;
+            return true;
         }
 
-        public void Remove(Tweet tweet)
+        public bool Remove(Tweet tweet)
         {
+            if (tweet == null)
+            {
+                return false;
+            }
             dbContext.Tweets.Remove(tweet);
             dbContext.SaveChanges();
+            return true;
         }
 
         public IList<Tweet> GetAll()
